requirements.txt
#aiofiles==24.1.0
#annotated-types==0.7.0
#anyio==4.5.0
carvekit==4.1.2
#certifi==2024.8.30
#charset-normalizer==3.3.2
#click==8.1.7
#colorama==0.4.6
#contourpy==1.3.0
#cycler==0.12.1
#fastapi==0.115.0
# filelock==3.16.1
# fonttools==4.53.1
# fsspec==2024.9.0
# h11==0.14.0
# huggingface-hub==0.25.2
# idna==3.10
# Jinja2==3.1.4
# joblib==1.4.2
# kiwisolver==1.4.7
# loguru==0.7.2
# MarkupSafe==2.1.5
matplotlib==3.9.2
# mpmath==1.3.0
# networkx==3.3
# npy-reader==1.1.2
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
# pyasn1==0.6.1
# pydantic==2.9.2
# pydantic_core==2.23.4
# pyparsing==3.1.4
# PyQt5==5.15.11
# PyQt5-Qt5==5.15.2
# PyQt5_sip==12.15.0
# PySimpleGUI==5.0.6
# python-dateutil==2.9.0.post0
# python-multipart==0.0.9
# pytz==2024.2
# PyYAML==6.0.2
# regex==2024.9.11
# requests==2.32.3
# rsa==4.9
# safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
# seaborn==0.13.2
# setuptools==75.1.0
# six==1.16.0
# sniffio==1.3.1
# starlette==0.38.5
# sympy==1.13.3
# threadpoolctl==3.5.0
timm==1.0.11
# tokenizers==0.20.1
torch==2.4.1
torchvision==0.19.1
# tqdm==4.66.5
transformers==4.45.2
# typing==3.7.4.3
# typing_extensions==4.12.2
# tzdata==2024.1
# urllib3==2.2.3
# uvicorn==0.30.6
# win32-setctime==1.1.0